FileUpload controller.

+++ UTB.EShop.Application/Constants/Constants.cs
@@ -2,8 +2,8 @@
 
 public static class Constants
 {
-    public const int MaxImageFileUpload = 1;
-    public const int MinImageFileUploadSize = 1_000;
-    public const int MaxImageFileUploadSize = 5_000_000;
+    public const long MaxImageFileUpload = 1;
+    public const long MinImageFileUploadSize = 1_000;
+    public const long MaxImageFileUploadSize = 5_000_000;
     public const int MaxPageSize = 10;
 }
